Major: Models added and performance compared
1. Added all models
2. Organized code
3. Added train_model.py and evaluate_model.py
4. Added Plot to compare performance

+++ data/get_dataset.py
@@ -1,15 +1,21 @@
 import pandas as pd
+import os
 
-url = "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data"
+def get_dataset():
+    if os.path.exists("adult_income.csv"):
+            print("Dataset already exists. Skipping download.\n")
+            return
 
-column_names = [
-    "age", "workclass", "fnlwgt", "education", "education-num", "marital-status",
-    "occupation", "relationship", "race", "sex", "capital-gain", "capital-loss",
-    "hours-per-week", "native-country", "income"
-]
+    url = "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data"
 
-df = pd.read_csv(url, names=column_names, na_values=" ?", skipinitialspace=True)
+    column_names = [
+        "age", "workclass", "fnlwgt", "education", "education-num", "marital-status",
+        "occupation", "relationship", "race", "sex", "capital-gain", "capital-loss",
+        "hours-per-week", "native-country", "income"
+    ]
 
-df.to_csv("adult_income.csv", index=False)
+    df = pd.read_csv(url, names=column_names, na_values=" ?", skipinitialspace=True)
 
-print("Dataset downloaded! Saved as 'adult_income.csv'")
+    df.to_csv("adult_income.csv", index=False)
+
+    print("Dataset downloaded! Saved as 'adult_income.csv'")
